requirements.txt
fastapi==0.109.1
uvicorn==0.23.2
python-multipart==0.0.19
lxml==4.9.3
pycryptodome==3.19.1
pydantic==2.4.2
python-dotenv==1.0.0
requests==2.32.3
beautifulsoup4==4.13.4
openai==1.78.0